refactor: separate resources into different files
