+++ src/resources/index.js
@@ -1,0 +1,3 @@
+export { createResource } from './resources'
+export { createDocumentResource } from './documentResource'
+export { createListResource } from './listResource'
